frontend: Refactor Improvement Pages to use a single generic component
- The pages were almost exactly the same, this way we remove a lot of code duplication.
